// src/domain/enums/college-course.ts
export enum CollegeCourse {
  MEDICINA = 'MEDICINA',
  ENGENHARIA = 'ENGENHARIA',
  DIREITO = 'DIREITO',
  ADMINISTRACAO = 'ADMINISTRACAO',
  CONTABILIDADE = 'CONTABILIDADE',
  PSICOLOGIA = 'PSICOLOGIA',
  PEDAGOGIA = 'PEDAGOGIA',
  ENFERMAGEM = 'ENFERMAGEM',
  FARMACIA = 'FARMACIA',
  FISIOTERAPIA = 'FISIOTERAPIA',
  ODONTOLOGIA = 'ODONTOLOGIA',
  VETERINARIA = 'VETERINARIA',
  ARQUITETURA = 'ARQUITETURA',
  CIENCIA_COMPUTACAO = 'CIENCIA_COMPUTACAO',
  SISTEMAS_INFORMACAO = 'SISTEMAS_INFORMACAO',
  JORNALISMO = 'JORNALISMO',
  PUBLICIDADE = 'PUBLICIDADE',
  MARKETING = 'MARKETING',
  ECONOMIA = 'ECONOMIA',
  RELACOES_INTERNACIONAIS = 'RELACOES_INTERNACIONAIS',
  OUTROS = 'OUTROS',
}